+++ internal/stackql/internal_data_transfer/builder_input/builder_input.go
@@ -43,6 +43,10 @@ type BuilderInput interface {
 	SetIsUndo(isUndo bool)
 	SetRequiredDataRequestKey(key string)
 	GetRequiredDataRequestKey() (string, bool)
+	SetPushdownQueryParams(queryParams map[string]string)
+	GetPushdownQueryParams() (map[string]string, bool)
+	SetPushdownLimit(limit int)
+	GetPushdownLimit() (int, bool)
 	SetDependencyNode(dependencyNode primitivegraph.PrimitiveNode)
 	SetParserNode(node sqlparser.SQLNode)
 	SetParamMap(paramMap map[int]map[string]interface{})
@@ -86,6 +90,9 @@ type builderInput struct {
 	tableInsertionContainer tableinsertioncontainer.TableInsertionContainer
 	insertCtx               drm.PreparedStatementCtx
 	requiredDataRequestKey  string
+	pushdownQueryParams     map[string]string
+	pushdownLimit           int
+	pushdownLimitSet        bool
 }
 
 func NewBuilderInput(
@@ -104,6 +111,27 @@ func NewBuilderInput(
 
 func (bi *builderInput) GetRequiredDataRequestKey() (string, bool) {
 	return bi.requiredDataRequestKey, bi.requiredDataRequestKey != ""
+}
+
+// SetPushdownQueryParams records the request query params computed by the analysis phase
+// for the acquire to push to the upstream API; the executor applies them verbatim.
+func (bi *builderInput) SetPushdownQueryParams(queryParams map[string]string) {
+	bi.pushdownQueryParams = queryParams
+}
+
+func (bi *builderInput) GetPushdownQueryParams() (map[string]string, bool) {
+	return bi.pushdownQueryParams, len(bi.pushdownQueryParams) > 0
+}
+
+// SetPushdownLimit records the LIMIT the analysis phase resolved as pushable (used by the
+// GraphQL acquire path to bound the page size); the acquire only carries it out.
+func (bi *builderInput) SetPushdownLimit(limit int) {
+	bi.pushdownLimit = limit
+	bi.pushdownLimitSet = true
+}
+
+func (bi *builderInput) GetPushdownLimit() (int, bool) {
+	return bi.pushdownLimit, bi.pushdownLimitSet
 }
 
 func (bi *builderInput) SetRequiredDataRequestKey(key string) {
@@ -300,5 +328,8 @@ func (bi *builderInput) Clone() BuilderInput {
 		isReturning:            bi.isReturning,
 		insertCtx:              bi.insertCtx,
 		requiredDataRequestKey: bi.requiredDataRequestKey,
+		pushdownQueryParams:    bi.pushdownQueryParams,
+		pushdownLimit:          bi.pushdownLimit,
+		pushdownLimitSet:       bi.pushdownLimitSet,
 	}
 }
